tp8/requirements.txt
streamlit
matplotlib
pandas
sklearn.linear_model
numpy